Modified Gender and Country Controller to have admin validation

+++ src/main/java/ittalents_final_project/ninegag/Controllers/GenderController.java
@@ -2,10 +2,14 @@ package ittalents_final_project.ninegag.Controllers;
 
 import ittalents_final_project.ninegag.Models.DAO.GenderDAOImplem;
 import ittalents_final_project.ninegag.Models.POJO.Gender;
+import ittalents_final_project.ninegag.Utils.Exceptions.NotLoggedException;
+import ittalents_final_project.ninegag.Utils.Exceptions.PermitionDeniedException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
+
+import javax.servlet.http.HttpSession;
 
 
 @RestController
@@ -15,14 +19,23 @@ public class GenderController extends BaseController {
     GenderDAOImplem dao;
 
 
-    @PostMapping(value = "/addGender")
-    public void addGender(@RequestBody Gender gender){
-        dao.addGender(gender);
+    @PostMapping(value = "/add/Gender")
+    public void addGender(@RequestBody Gender gender, HttpSession session) throws NotLoggedException, PermitionDeniedException{
+        if (validateAdmin(session)) {
+            dao.addGender(gender);
+        } else {
+            throw new PermitionDeniedException("You are not admin");
+        }
     }
 
-    @PostMapping(value = "/deleteGender")
-    public void removeGender(@RequestBody Gender gender){
-        dao.deleteGenderByID(gender.getGender_ID());
+    @PostMapping(value = "/delete/Gender")
+    public void removeGender(@RequestBody Gender gender, HttpSession session) throws NotLoggedException, PermitionDeniedException{
+        if (validateAdmin(session)) {
+            dao.deleteGenderByID(gender.getGender_ID());
+        } else {
+            throw new PermitionDeniedException("You are not admin");
+        }
+
     }
 
 }
